backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
supabase==2.8.0
psycopg[binary]==3.2.3
google-generativeai==0.8.3
pymupdf==1.24.10
numpy==2.1.1
python-multipart==0.0.12

